v-xent/requirements.txt
shodan
virustotal-api
python-dotenv
ratelimit
requests
colorama
